chore(deploy): don't install object-cache on testing env

+++ config/post-deploy.sh
@@ -31,18 +31,21 @@ if [ -z $WP_HOME ]; then
     exit 1
 fi
 
-if [ ! -f wp-cli.local.yml ]; then
+if [ 'testing' != "$WP_ENV" ] && [ ! -f wp-cli.local.yml ]; then
     cp wp-cli.yml wp-cli.local.yml
     sed -i -E "s~;url =.*~url = ${WP_HOME}~" wp-cli.local.yml
     _suc 'File: `wp-cli.local.yml` created successfully'
-else
-    _inf 'File: `wp-cli.local.yml` already present'
 fi
 
 if [ -f .env ]; then
-    vendor/bin/wp dotenv list
+    vendor/bin/wp dotenv salts generate
+    vendor/bin/wp dotenv set WP_ENV "$WP_ENV"
+    vendor/bin/wp dotenv set DB_HOST "$DB_HOST"
+    vendor/bin/wp dotenv set WP_HOME "$WP_HOME"
+    vendor/bin/wp dotenv list --format=table
 fi
 
+vendor/bin/wp --info
 vendor/bin/wp core install --skip-email --title="WordPress Site" \
     --admin_user="admin" --admin_password="secret" --admin_email="admin@example.com"
 
@@ -51,7 +54,7 @@ vendor/bin/wp option update permalink_structure '/%postname%/'
 
 vendor/bin/wp cache flush
 
-if [ ! -f public/app/object-cache.php ] && [ -f public/app/mu-plugins/redis-cache/includes/object-cache.php ]; then
+if [ 'testing' != "$WP_ENV" ] && [ ! -f public/app/object-cache.php ] && [ -f public/app/mu-plugins/redis-cache/includes/object-cache.php ]; then
     cp public/app/mu-plugins/redis-cache/includes/object-cache.php public/app/object-cache.php
     _suc 'Drop-in Object-Cache instaled successfully'
 fi
